feat: implement service layers
