+++ app/src/main/java/br/com/fsdev/learningapp/data/mappers/CharacterMapper.kt
@@ -1,8 +1,8 @@
 package br.com.fsdev.learningapp.data.mappers
 
 import br.com.fsdev.learningapp.data.repository.CharacterResponse
-import br.com.fsdev.learningapp.domain.Character
-import br.com.fsdev.learningapp.domain.CharacterStatus
+import br.com.fsdev.learningapp.domain.models.Character
+import br.com.fsdev.learningapp.domain.models.CharacterStatus
 
 object CharacterMapper {
     fun toDomain(response: CharacterResponse) =
@@ -11,8 +11,7 @@ object CharacterMapper {
                 id = id,
                 name = name,
                 status = CharacterStatus.toStatus(status),
-                species = species,
-                origin = origin
+                species = species
             )
         }
 }
