+++ app/src/main/java/br/com/fsdev/learningapp/data/repository/CharacterResponse.kt
@@ -12,6 +12,5 @@ data class CharacterResponse(
     val id: Int,
     val name: String,
     val status: String,
-    val species: String,
-    val origin: String
+    val species: String
 )
