+++ app/src/main/java/br/com/fsdev/learningapp/data/repository/RickAndMortyGateway.kt
@@ -3,7 +3,7 @@ package br.com.fsdev.learningapp.data.repository
 import retrofit2.http.GET
 import retrofit2.http.Path
 
-interface CharacterGateway {
+interface RickAndMortyGateway {
 
     @GET("character")
     suspend fun getCharacters() : CharactersResponse
