+++ app/src/main/java/br/com/fsdev/learningapp/domain/models/Character.kt
@@ -1,0 +1,10 @@
+package br.com.fsdev.learningapp.domain.models
+
+import java.io.Serializable
+
+data class Character(
+    val id: Int,
+    val name: String,
+    val status: CharacterStatus,
+    val species: String
+): Serializable
